+++ .php-cs-fixer.dist.php
@@ -1,7 +1,9 @@
 <?php
 
 $finder = PhpCsFixer\Finder::create()
-    ->in(__DIR__)
+    ->in(__DIR__.'/examples')
+    ->in(__DIR__.'/src')
+    ->in(__DIR__.'/tests')
 ;
 
 $config = new PhpCsFixer\Config();
@@ -10,6 +12,13 @@ return $config
     ->setRules([
         '@Symfony' => true,
         'declare_strict_types' => true,
+        'global_namespace_import' => [
+            'import_constants' => true,
+            'import_functions' => true,
+            'import_classes' => true,
+        ],
+        'yoda_style' => false,
+        'increment_style' => ['style' => 'post'],
     ])
     ->setFinder($finder)
 ;