+++ src/Ip.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace Flow;
 
-final class Ip
+readonly final class Ip
 {
-    public function __construct(public readonly ?object $data = null)
+    public function __construct(public ?object $data = null)
     {
     }
 }