+++ tests/Driver/DriverTest.php
@@ -5,18 +5,18 @@ declare(strict_types=1);
 namespace Flow\Test\Driver;
 
 use Exception;
-use PHPUnit\Framework\TestCase;
 use Flow\DriverInterface;
+use PHPUnit\Framework\TestCase;
 use Throwable;
 
 abstract class DriverTest extends TestCase
 {
     abstract protected function createDriver(): DriverInterface;
 
-    public function testCoroutine(): void
+    public function testAsync(): void
     {
         $driver = $this->createDriver();
-        $driver->coroutine(static function () {
+        $driver->async(static function () {
         }, function (?Throwable $e) use ($driver) {
             $driver->stop();
             $this->assertNull($e);
@@ -24,14 +24,26 @@ abstract class DriverTest extends TestCase
         $driver->start();
     }
 
-    public function testCoroutineError(): void
+    public function testAsyncError(): void
     {
         $driver = $this->createDriver();
-        $driver->coroutine(static function () {
+        $driver->async(static function () {
             throw new Exception();
         }, function (?Throwable $e) use ($driver) {
             $driver->stop();
             $this->assertNotNull($e);
+        })();
+        $driver->start();
+    }
+
+    public function testDelay(): void
+    {
+        $driver = $this->createDriver();
+        $driver->async(static function () use ($driver) {
+            $driver->delay(1 / 1000);
+        }, function (?Throwable $e) use ($driver) {
+            $driver->stop();
+            $this->assertNull($e);
         })();
         $driver->start();
     }
@@ -41,10 +53,10 @@ abstract class DriverTest extends TestCase
         $i = 0;
         $driver = $this->createDriver();
         $driver->tick(1, function () use (&$i) {
-            ++$i;
+            $i++;
         });
-        $driver->coroutine(function () use ($driver, &$i) {
-            usleep(3000);
+        $driver->async(function () use ($driver, &$i) {
+            $driver->delay(3 / 1000);
             $driver->stop();
 
             $this->assertGreaterThan(3, $i);