+++ tests/Flow/FlowTest.php
@@ -7,22 +7,25 @@ namespace Flow\Test\Flow;
 use ArrayObject;
 use Closure;
 use Flow\DriverInterface;
+use Flow\Flow\Flow;
 use Flow\Ip;
 use Flow\IpStrategyInterface;
-use Flow\Flow\Flow;
 use RuntimeException;
-use Throwable;
 
 class FlowTest extends AbstractFlowTest
 {
     /**
      * @dataProvider jobProvider
+     *
+     * @param array<Closure> $jobs
      */
-    public function testJob(DriverInterface $driver, IpStrategyInterface $ipStrategy, Closure $job, int $resultNumber): void
+    public function testJob(DriverInterface $driver, IpStrategyInterface $ipStrategy, array $jobs, int $resultNumber): void
     {
         $ip = new Ip(new ArrayObject(['number' => 0]));
-        $errorJob = function() {};
-        $flow = new Flow($job, $errorJob, $ipStrategy, $driver);
+        $flow = array_reduce(
+            array_map(fn ($job) => new Flow($job, static function () {}, $ipStrategy, $driver), $jobs),
+            fn ($flow, $flowIt) => $flow ? $flow->fn($flowIt) : $flowIt
+        );
         ($flow)($ip, function (Ip $ip) use ($driver, $resultNumber) {
             $driver->stop();
             self::assertSame(ArrayObject::class, $ip->data::class);
@@ -45,10 +48,8 @@ class FlowTest extends AbstractFlowTest
         }, function (object $data): void {
             $data['n2'] *= 4;
         }];
-        $errorJobs = [function() {
-
-        }, function() {
-
+        $errorJobs = [function () {
+        }, function () {
         }];
         $flow = new Flow($jobs, $errorJobs, null, $driver);
 
@@ -56,7 +57,7 @@ class FlowTest extends AbstractFlowTest
 
         $callback = function (Ip $ip) use ($driver, &$ips, $ip1, $ip2) {
             $ips[] = $ip;
-            if(count($ips) === 2) {
+            if (count($ips) === 2) {
                 $this->assertSame($ip1, $ips[0]);
                 $this->assertSame($ip2, $ips[1]);
                 self::assertSame(6, $ip1->data['n1']);
@@ -81,13 +82,24 @@ class FlowTest extends AbstractFlowTest
     {
         $exception = new RuntimeException('job error');
 
-        return $this->matrix([
-            'job' => [static function (ArrayObject $data) {
+        return $this->matrix(fn (DriverInterface $driver) => [
+            'oneJob' => [[static function (ArrayObject $data) {
                 $data['number'] = 5;
-            }, 5],
-            'exceptionJob' => [static function () use ($exception) {
+            }], 5],
+            'oneAsyncJob' => [[static function (ArrayObject $data) use ($driver): void {
+                $driver->delay(1 / 1000);
+                $data['number'] = 5;
+            }], 5],
+            'twoAsyncJob' => [[static function (ArrayObject $data) use ($driver): void {
+                $driver->delay(1 / 1000);
+                $data['number'] += 5;
+            }, static function (ArrayObject $data) use ($driver): void {
+                $driver->delay(1 / 1000);
+                $data['number'] *= 2;
+            }], 10],
+            'exceptionJob' => [[static function () use ($exception) {
                 throw $exception;
-            }, 0],
+            }], 0],
         ]);
     }
 }